feat: Add migration for adding "description" column to "categories" table
Add a new migration file to add a "description" column to the "categories" table in the database. This will allow storing descriptions for each category.

feat: Implement DeleteCategoryUseCase

Create the DeleteCategoryUseCase class in the use-cases folder of the category application module. This use case is responsible for deleting a category from the database.

refactor: Remove FindCategoriesUseCase

Remove the FindCategoriesUseCase class from the use-cases folder of the category application module. This use case is no longer needed.

feat: Add migration for adding "deleted_at" and "is_active" columns to multiple tables

Add new migration files to add "deleted_at" and "is_active" columns to the "categories", "installments", and "transactions" tables in the database. These columns will be used for soft deletion and to track the active status of records.

feat: Implement GetCategoryUseCase

Create the GetCategoryUseCase class in the use-cases folder of the category application module. This use case is responsible for retrieving a single category from the database based on its ID.

feat: Implement GetAllCategoriesUseCase

Create the GetAllCategoriesUseCase class in the use-cases folder of the category application module. This use case is responsible for retrieving all categories from the database.

feat: Implement GetCategoryController

Create the GetCategoryController class in the controllers folder of the category HTTP module. This controller handles the GET request to retrieve a single category based on its ID.

feat: Implement GetAllCategoriesController

Create the GetAllCategoriesController class in the controllers folder of the category HTTP module. This controller handles the GET request to retrieve all categories.

feat: Implement DeleteCategoryController

Create the DeleteCategoryController class in the controllers folder of the category HTTP module. This controller handles the DELETE request to delete a category based on its ID.

feat: Implement UpdateCategoryUseCase

Create the UpdateCategoryUseCase class in the use-cases folder of the category application module. This use case is responsible for updating a category in the database.

feat: Implement UpdateCategoryController

Create the UpdateCategoryController class in the controllers folder of the category HTTP module. This controller handles the PATCH request to update a category based on its ID.

feat: Implement CreateCategoryController

Create the CreateCategoryController class in the controllers folder of the category HTTP module. This controller handles the POST request to create a new category.

refactor: Remove CreateCategoryUseCase

Remove the CreateCategoryUseCase class from the use-cases folder of the category application module. This use case is no longer needed.